although this is really size, it needs to be able to hold -1 to indicate no valid popups exist either before the collection has been used or when a click had no results

+++ Shared/IdentifyController.h
@@ -71,7 +71,7 @@ private:
   bool canPrev() const;
 
   double m_tolerance = 5.0;
-  size_t m_currentPopup = -1;
+  int m_currentPopup = -1;
   std::vector<std::unique_ptr<Esri::ArcGISRuntime::Popup>> m_popups;
 };
